feat(screenshots): add screenshots for local and prod envs
This commit adds screenshots for the portfolio application for both the
local and production environments.

The screenshots are taken for the following pages:
- homepage
- about
- talks (card and list view)
- articles (card and list view)
- next-talks
- config
- sample talk
- sample article

The screenshots are stored in the  directory, with the
production screenshots in the  subdirectory.

A  file has been added to each directory to provide a list
of the screenshots.